# quarkus-order-service/target/classes/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: |
    Order management service. Handles order placement with:
    - Stock reservation (via Product Service)
    - Saga pattern compensation (stock restore on failure)
    - Outbox pattern for event persistence
    - Async notification dispatch (via CompletableFuture to Notification Service)
    Depends on: Cart Service, Product Service, Notification Service
  version: 1.0.0
  contact:
    name: Ecommerce Team
  x-service-port: 8084
  x-database: order_db
  x-framework: Quarkus
  x-calls:
    - Cart Service (GET /api/cart, DELETE /api/cart)
    - Product Service (PUT /api/products/{id}/reduce-stock, PUT /api/products/{id}/restore-stock)
    - Notification Service (POST /api/notifications/eventgrid)
  x-messaging: Reactive Messaging emitter on channel "order-placed" (Kafka-ready)
servers:
  - url: http://localhost:8084
    description: Local development
  - url: http://quarkus-order-service:8084
    description: Docker internal
security:
  - bearerAuth: []
tags:
  - name: Orders
    description: Order placement and retrieval
paths:
  /api/orders:
    get:
      summary: List customer orders
      description: Retrieves all orders for the authenticated customer.
      operationId: myOrders
      tags:
        - Orders
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            format: bearer
          description: JWT Bearer token from Auth Service
      responses:
        '200':
          description: List of orders for the customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Place a new order (checkout)
      description: |
        Places an order by executing a distributed transaction:

        1. **Fetch cart** — GET /api/cart from Cart Service (forwards JWT)
        2. **Reserve stock** — For each cart item, PUT /api/products/{id}/reduce-stock to Product Service
           - If any item fails stock reduction, **compensate** by restoring previously reduced stock (Saga)
        3. **Persist order** — Save Order document to MongoDB (order_db)
           - If persistence fails, **compensate** stock reductions (Saga)
        4. **Clear cart** — DELETE /api/cart from Cart Service (best-effort)
        5. **Outbox** — Save OrderPlacedEvent to outbox_events collection (for eventual message broker relay)
        6. **Notify** — Dispatch async HTTP POST to Notification Service (CompletableFuture.runAsync)

        The Authorization header is forwarded to Cart, Product, and Notification services to propagate the caller's identity.
      operationId: checkout
      tags:
        - Orders
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Cannot place an order with an empty cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}:
    get:
      summary: Get order by ID
      description: Retrieves a specific order by ID. The customer must own the order.
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID (MongoDB ObjectId)
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not your order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
          example: 64b7f3c2a8d2c910e4b5a3ea
        customerId:
          type: string
          example: 64b7f1e2a8d2c910e4b5a3d7
        customerEmail:
          type: string
          format: email
          example: customer@example.com
        customerName:
          type: string
          example: John Doe
        customerPhone:
          type: string
          example: "+1-555-0123"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
          example: 59.98
        status:
          type: string
          enum: [PLACED, CANCELLED, SHIPPED, DELIVERED, REFUNDED]
          example: PLACED
        createdAt:
          type: string
          format: date-time
          example: 2024-07-16T12:00:00Z
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          example: 64b7f1e2a8d2c910e4b5a3d8
        productName:
          type: string
          example: Wireless Headphones
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: double
          example: 29.99
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Cannot place an order with an empty cart
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time